OCEProject 3813 - Added MaximumRetries and ExposeRawResponse to make sure we can accurately capture errors for Search when any of the 3 nodes is down.
git-svn-id: https://ncisvn.nci.nih.gov/svn/oce_wcmteam/ContentDeliveryEngine/branches/Griffin@14543 e9cc8610-1bbf-714c-a5f3-26df0ede829a

+++ CDESites/DCEG/SiteSpecific/DCEG.Web/SnippetTemplates/NewsletterSearchResults.ascx.cs
@@ -145,7 +145,7 @@ namespace DCEG.Web.SnippetTemplates
                     if (results != null && results.ResultCount > 0)
                     {
 
-                        lastRecord = firstRecord + results.ResultCount - 1;
+                        lastRecord = firstRecord + (int)results.ResultCount - 1;
                         totalItems = (int)results.ResultCount;
 
                         ResultRepeater.DataSource = results;
